02/about (#11)
* feat: about component

Created the About components on components/.

* style: changed the About component to fit the requirements in pull request 11

I changed the Analytics section to make all numbers reach the maximum at the same time, changed the projects for the top 3 in Devpost and made them a showcase instead of a carousel. Removed the black background of the section because setting it to max width so that it looks consistent would conflict with global styles. Centered the paragraphs.

* refac: column format

* fix: correct some spelling mistakes

---------

Co-authored-by: Jeffrey Zang <zangj4548@wrdsb.ca>
Co-authored-by: Jeffrey Zang <66485719+jeffrey-zang@users.noreply.github.com>